Mini_project-master/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
pydantic
networkx
pandas
openpyxl
xlrd<2.0
statsmodels
python-multipart
aiofiles
redis
celery
boto3
python-dotenv
pytest
pytest-asyncio
langchain_experimental
langchain-groq

tabulate